feat(socket): 增加订阅主题 ACL 鉴权策略和优雅停机支持
- 新增 DefaultTopicAclPolicy 实现，支持管理员保护区、用户私有前缀及租户隔离校验
- StompMessageInterceptor 支持基于 TopicAclPolicy 的 SUBSCRIBE ACL 鉴权，拒绝时返回 ERROR 帧
- WebSocketAuthInterceptor 增加优雅停机检查，拒绝停机期间新握手请求，返回 503
- WebSocketAutoConfiguration 自动配置默认 TopicAclPolicy Bean 和优雅停机相关依赖注入
- WebSocketSessionAttributes 使用统一常量定义 Session 属性键值，避免硬编码
- 代码增强日志警告，拒绝未认证用户订阅及 ACL 拒绝情况记录日志
- 完善单元测试覆盖 PiiContextExtractor 上下文提取逻辑验证
- 统一调整搜索缓存服务对 Redis 的访问方式及序列化实现，提升容错性
- Seata 模块添加 undo_log 表架构启动时校验功能，避免运行时异常
- PII 检测器增强 API Key、IP、护照号、JWT 等类型的上下文信息提取和记录

+++ ydsz-common/ydsz-common-search/src/main/java/com/njydsz/common/search/service/UnifiedSearchService.java
@@ -395,7 +395,7 @@ public class UnifiedSearchService {
    *
    * @return 缓存中的搜索响应条目数量
    */
-  public int getCacheSize() {
+  public long getCacheSize() {
     return cacheService.size();
   }
